+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@ changelog, see [CHANGELOG.md](CHANGELOG.md).
 
 ### Fixed
 
+- **"Refresh cache" option missing on combined shelves with online sources.** The action only showed up for pure wishlist / store shelves — picking a composite shelf that included a wishlist or store source from the card menu had no way to clear the cached results. Now the option is back in the Shelf → Management submenu whenever any online source is involved.
+- **NEW / discount badge no longer disappears or duplicates on the focused card.** A recent change removed the dedicated overlay that drew the badge above Steam's focus ring; the in-card badge alone is painted under the ring, so the badge looked like it had vanished the moment a card was selected. Bringing the overlay back without coordination left both copies stacking visibly. The overlay is back AND the inline copy is hidden on focus, so exactly one badge shows in every state.
 - **Home lag (#81).** Each card with a "NEW" or discount badge was running its own DOM observers and focus listeners to track the QAM/modal overlay state. On a home with 30+ visible cards, that turned into dozens of observers all reacting to every DOM change. Now a single shared detector serves the whole home — idle CPU drops noticeably.
 - **Random-sorted shelves now actually re-shuffle (#82).** The previous build cached the shuffle for 24 hours, and the cache survived Steam restarts — so a shelf with random sort would stay frozen in the same order until the next day. Now the cache is wiped at every plugin boot and on every shelf refresh.
 - **View button on running games no longer shows "Application already open".** Pressing View on a game that's running now correctly returns you to the game without the error toast — same behaviour as picking the menu's first item manually.
