+++ src/components/HomeInject.tsx
@@ -2,7 +2,6 @@ import { useEffect, useMemo, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { createPortal } from "react-dom";
 import { ShelfView } from "./Shelf";
-import { BadgeFocusOverlay } from "./shelf/BadgeFocusOverlay";
 import type { Settings, Shelf, SmartShelf, SmartShelfMode } from "../types";
 import { refreshSettings, subscribeSettings, saveSettings, getCurrentSettings } from "../settingsStore";
 import { useContainerDragReorder } from "../core/reorder";
@@ -11,6 +10,7 @@ import { createDeckyPlatform } from "../runtime/deckyPlatform";
 import { logInfo, logWarn } from "../runtime/logger";
 import { logDiagnostic } from "../runtime/diagnostics";
 import { getPreferredSteamDocument, getPreferredSteamWindow, getAllSteamDocuments } from "../runtime/steamHost";
+import { ROOT_ID, seededShuffle, isHomeRoute, hasHomeDomSignals, detectNavTreeApi, findOrCreateMount } from "./home/mountUtils";
 import { applyHideRecents, reapplyHomeHides, applyHideHomeTabs, getMountFailed } from "../runtime/homePatch";
 import { getRecentsReplaceFailed, subscribeRecentsReplaceFailed, isRecentsReplaceInjecting, subscribeRecentsReplaceInjecting, getRecentsReplaceActiveShelfId } from "../runtime/recentsReplace";
 import { Focusable } from "../runtime/host/decky";
@@ -23,13 +23,8 @@ import { isInVisibilityWindow, nextVisibilityBoundary, getModeVisibilityWindows,
 import { flowChildrenProps } from "../core/steamOSVersion";
 import { getRuntimeClassMap } from "../core/webpackCompat";
 import { isCssLoaderActive, getNativeRecentsClassName, isArtHeroActive, isNoHeroGradientActive, isHeroFullscreenActive, isNoHomeTextActive, isFocusRoundCompatActive, isTiltedHomeActive, getTiltedHomeMode } from "../core/cssLoaderDetect";
+import { BadgeFocusOverlay } from "./shelf/BadgeFocusOverlay";
 
-// Fallback for the native shelf-section token when the runtime classmap
-// hasn't been populated yet. Mirrors `FALLBACK_SHELF_SECTION` in
-// `runtime/homePatch.tsx`.
-const FALLBACK_SHELF_SECTION = "_282X0J4BtrSF1IXctmOe-X";
-
-const ROOT_ID = "deck-shelves-home-root";
 const homePlatform = createDeckyPlatform();
 
 const SURPRISE_MODES: SmartShelfMode[] = [
@@ -38,196 +33,7 @@ const SURPRISE_MODES: SmartShelfMode[] = [
   "non_steam", "spare_time", "time_of_day", "rediscover", "forgotten",
 ];
 
-function seededShuffle<T>(arr: T[], seed: number): T[] {
-  const out = [...arr];
-  let s = (seed | 0) >>> 0;
-  for (let i = out.length - 1; i > 0; i--) {
-    s = (Math.imul(s, 1664525) + 1013904223) >>> 0;
-    const j = s % (i + 1);
-    [out[i], out[j]] = [out[j], out[i]];
-  }
-  return out;
-}
-
-
-// Navigation patches (reparentNavTreeNodes, patchMenuButton, patchShelfEdgeNavigation)
-// are in ./home/navPatches.ts
-
-function isHomeRoute(): boolean {
-  // Steam Deck 3.9: the BigPicture window stays at `/index.html` regardless
-  // of the React Router route, while the SharedJSContext window carries the
-  // real `/routes/library/home` path. Inspecting only the preferred window
-  // would lock the result to one of those two truths. Walk every known Steam
-  // window so we catch whichever one is currently tracking the route.
-  const wins: Window[] = [];
-  try { wins.push(getPreferredSteamWindow()); } catch {}
-  try { for (const d of getAllSteamDocuments()) { const w = d.defaultView; if (w && !wins.includes(w)) wins.push(w); } } catch {}
-  for (const win of wins) {
-    try {
-      const href = `${win.location?.pathname ?? ""}${win.location?.hash ?? ""}`.toLowerCase();
-      // Strict: only the actual home route (`/routes/library/home`) counts.
-      // The previous permissive "any /library that isn't /app/ or /collections"
-      // check matched library tabs (e.g. `/routes/library/games`), so leaving
-      // home for the library tab kept `wasOnHome` true — and the B-return
-      // never crossed the `nowOnHome && !wasOnHome` edge, so the focus
-      // restore never fired.
-      if (href.includes("/library/home")) return true;
-    } catch {}
-  }
-  return false;
-}
-
-function hasHomeDomSignals(): boolean {
-  const doc = getPreferredSteamDocument();
-  if (!doc) return false;
-  if (doc.querySelector('[class*="libraryhome"], [class*="LibraryHome"], [class*="BasicHomeView"], [class*="gamepadlibrary"]')) return true;
-  if (doc.querySelector('[aria-label="Jogos recentes"], [aria-label="Recent Games"], [class*="ReactVirtualized__Grid"][aria-label]')) return true;
-  try {
-    const token = getRuntimeClassMap(doc)?.shelfSection || FALLBACK_SHELF_SECTION;
-    if (doc.querySelector(`div.${token}, [class*="${token}"]`)) return true;
-  } catch (e) { logInfo("HOME", "hasHomeDomSignals: class selector failed", String(e)); }
-  return false;
-}
-
-function detectNavTreeApi(): { available: boolean; detail: string } {
-  try {
-    const ctrl = (globalThis as any).FocusNavController
-      ?? (globalThis as any).GamepadNavTree?.m_context?.m_controller;
-    if (!ctrl) return { available: false, detail: 'no FocusNavController' };
-    const ctx = ctrl.m_ActiveContext || ctrl.m_LastActiveContext;
-    const trees: any[] = ctx?.m_rgGamepadNavigationTrees ?? [];
-    const main = trees.find((t: any) => t.m_ID === "GamepadUI_Full_Root");
-    if (!main) return { available: false, detail: 'no GamepadUI_Full_Root tree' };
-    const root = main.Root || main.m_Root;
-    if (!root) return { available: false, detail: 'no Root on main tree' };
-    if (!Array.isArray(root.m_rgChildren)) return { available: false, detail: 'm_rgChildren unavailable' };
-    return { available: true, detail: `${root.m_rgChildren.length} root children` };
-  } catch (e) {
-    return { available: false, detail: String(e) };
-  }
-}
-
-function resolveAnchor(doc?: Document): { parent: HTMLElement; before: ChildNode | null } | null {
-  doc = doc ?? getPreferredSteamDocument();
-  if (!doc) return null;
-
-  // Strategy: find the "Recent Games" section, then walk UP to the scrollable
-  // viewport and insert as a direct child AFTER the Recent Games chain.
-  // This prevents our mount from expanding the native section and overlapping
-  // subsequent native content (e.g., "What's New" tabs).
-  const recentLabels = ["jogos recentes", "recent games", "recently played", "jogados recentemente"];
-  const candidates = Array.from(doc.querySelectorAll('[role="list"],[aria-label],[class*="ReactVirtualized__Grid"]'));
-  for (const node of candidates) {
-    const txt = `${(node.getAttribute?.("aria-label") || "")} ${(node.textContent || "")}`.toLowerCase();
-    if (!recentLabels.some((l) => txt.includes(l))) continue;
-    // Walk up to the scrollable viewport ancestor
-    let container: HTMLElement | null = node as HTMLElement;
-    for (let i = 0; i < 12 && container; i++) {
-      const p: HTMLElement | null = container.parentElement;
-      if (!p || p === doc.body) break;
-      try {
-        // Use the element's own window for getComputedStyle so cross-doc
-        // anchoring (BigPicture vs SharedJSContext) still resolves overflow.
-        const win = p.ownerDocument?.defaultView ?? getPreferredSteamWindow();
-        const cs = win.getComputedStyle(p);
-        const oy = (cs.overflowY || '').toLowerCase();
-        if ((oy === 'auto' || oy === 'scroll') && p.scrollHeight > p.clientHeight) {
-          // Found the scrollable viewport — insert after the current container
-          return { parent: p, before: container.nextSibling };
-        }
-      } catch (e) { logInfo("HOME", "resolveAnchor: getComputedStyle failed", String(e)); }
-      container = p;
-    }
-    // Fallback: find first ancestor with multiple children
-    container = node as HTMLElement;
-    for (let i = 0; i < 6 && container; i++) {
-      const p: HTMLElement | null = container.parentElement;
-      if (!p || p === doc.body) break;
-      if (p.childElementCount > 1) {
-        return { parent: p, before: container.nextSibling };
-      }
-      container = p;
-    }
-  }
-
-  const chipLabels = ["what's new", "friends", "recommended", "novidades", "amigos", "recomendados"];
-  for (const node of Array.from(doc.querySelectorAll('button, [role="tab"]'))) {
-    const text = (node.textContent || "").trim().toLowerCase();
-    if (!chipLabels.includes(text)) continue;
-    let row: HTMLElement | null = node.parentElement as HTMLElement;
-    while (row && row.childElementCount <= 1 && row !== doc.body) row = row.parentElement;
-    if (row?.parentElement && row !== doc.body) return { parent: row.parentElement, before: row };
-  }
-
-  const containers = Array.from(doc.querySelectorAll('[class*="gamepadlibrary"],[class*="libraryhome"],[class*="BasicHomeView"],main,[role="main"]'));
-  for (const node of containers) {
-    if (node instanceof HTMLElement) return { parent: node, before: node.firstChild };
-  }
-
-  return null;
-}
-
-// Holds the last mount we created so we can re-insert THE SAME element back
-// into the DOM after Steam blows it away during navigation. Reusing the same
-// node keeps React's portal target stable — no unmount/remount of children,
-// so per-shelf data, focus state, nav-tree registration, and ArtHero
-// data-attribute all survive the brief detach.
-let lastCreatedMount: HTMLElement | null = null;
-
-function findOrCreateMount(): HTMLElement | null {
-  // SteamOS 3.9: `preferredSteamWindow` can get stuck on the SharedJSContext
-  // while the visual DOM lives in the BigPicture main window. Sweeping every
-  // known Steam doc — preferred first for 3.7 fast-path parity — guarantees
-  // we re-attach to whichever doc actually carries the home anchor when our
-  // mount gets blown away by Steam re-renders.
-  const preferred = getPreferredSteamDocument();
-  const allDocs = getAllSteamDocuments();
-  const seen = new Set<Document>();
-  const docs: Document[] = [];
-  const push = (d: Document | null | undefined) => {
-    if (d && !seen.has(d)) { seen.add(d); docs.push(d); }
-  };
-  push(preferred);
-  for (const d of allDocs) push(d);
-
-  // 1) Reuse a still-connected mount in any known doc (covers cases where
-  //    the mount survives in BigPic but preferred points at SWC).
-  for (const d of docs) {
-    const existing = d.getElementById(ROOT_ID) as HTMLElement | null;
-    if (existing?.isConnected) return existing;
-  }
-
-  // 2) Mount was detached but element ref is still alive — re-insert THE SAME
-  //    element into the first doc with a valid anchor. React's portal keeps
-  //    rendering into it; the brief detach is invisible to the React subtree.
-  if (lastCreatedMount && !lastCreatedMount.isConnected) {
-    for (const d of docs) {
-      const anchor = resolveAnchor(d);
-      if (!anchor || anchor.parent === d.body) continue;
-      // If the cached element belongs to a different doc, adoptNode keeps it
-      // usable in the new doc — Steam doc instances share a window in 3.7 so
-      // this is normally a no-op; in 3.9 the cross-doc case is harmless.
-      try { if (lastCreatedMount.ownerDocument !== d) d.adoptNode(lastCreatedMount); } catch {}
-      anchor.parent.insertBefore(lastCreatedMount, anchor.before);
-      return lastCreatedMount;
-    }
-  }
-
-  // 3) Otherwise create a new mount in the first doc whose `resolveAnchor`
-  //    yields a usable insertion point.
-  for (const d of docs) {
-    const anchor = resolveAnchor(d);
-    if (!anchor || anchor.parent === d.body) continue;
-    const mount = d.createElement("div");
-    mount.id = ROOT_ID;
-    mount.style.cssText = "width:100%;display:block;position:relative;z-index:0;margin:0;padding:0;";
-    anchor.parent.insertBefore(mount, anchor.before);
-    lastCreatedMount = mount;
-    logInfo("HOME", "mount created", { parent: anchor.parent.tagName });
-    return mount;
-  }
-  return null;
-}
+// Mount + anchor + home-detection helpers live in ./home/mountUtils.
 
 export function HomeShelves() {
   const { t } = useTranslation();
