+++ src/core/steamGameMenu.ts
@@ -16,6 +16,22 @@ import { saveFocusTarget, beginFocusRestoreLoop } from "./focusRestore";
 import { invalidateRandomSortCache } from "../steam";
 import { invalidateSmartShelfCache } from "../steam/smartShelves";
 import { triggerShelfRefresh } from "./shelfRefresh";
+import { isOnlineSource } from "../domain/sourceUtils";
+import {
+  buildDeckShelvesMenuItems as buildDeckShelvesMenuItemsBase,
+  buildShelfContextMenu,
+  buildLibraryAddToShelfItems,
+} from "./menu/itemBuilders";
+
+export { buildShelfContextMenu, buildLibraryAddToShelfItems };
+
+// Thin wrapper around the extracted builder — injects the module-global
+// focused-card context (`_activeAppIdForMenu` / `_activeCardIndexForMenu`)
+// so existing call sites don't have to forward those values. Signature
+// matches the pre-extraction local function.
+function buildDeckShelvesMenuItems(shelfId: string, dfl: any, R: any, appid?: number): any[] {
+  return buildDeckShelvesMenuItemsBase(shelfId, dfl, R, appid, _activeAppIdForMenu, _activeCardIndexForMenu);
+}
 
 /**
  * Returns `true` when this device should use the pre-3.8 (v1.4.0-style) menu
@@ -643,317 +659,6 @@ export function installCreateContextMenuPatch(): void {
     };
     _ctxMenuPatchInstalled = true;
   } catch {}
-}
-
-/**
- * Builds a `Deck Shelves` nested submenu (MenuGroup) containing the seven
- * per-shelf actions: Edit, Duplicate, Collapse/Expand, Hide/Show, Move up,
- * Move down, Delete. Returned as a single MenuGroup so the parent menu
- * shows "Deck Shelves" as one entry that expands into the action list.
- *
- * Edit and Delete navigate to routes that mount a standalone settings
- * controller and open the modal via DFL.showModal — no QAM dependency.
- * The other actions call the standalone helpers in `shelfActions.ts`.
- *
- * Returns an empty array when the shelf cannot be located (silent no-op
- * — the native menu renders unchanged).
- */
-function buildDeckShelvesMenuItems(shelfId: string, dfl: any, R: any, appid?: number): any[] {
-  const focusedAppId = appid ?? _activeAppIdForMenu;
-  if (!dfl?.MenuItem || !dfl?.MenuGroup || !R?.createElement) return [];
-  const settings = getCurrentSettings?.();
-  if (!settings) return [];
-  const shelves = settings.shelves ?? [];
-  const smartShelves = (settings as any).smartShelves ?? [];
-  let idx = shelves.findIndex((sh: any) => sh.id === shelfId);
-  let isSmart = false;
-  let listLen = shelves.length;
-  let shelf: any = idx >= 0 ? shelves[idx] : null;
-  if (!shelf) {
-    idx = smartShelves.findIndex((sh: any) => sh.id === shelfId);
-    if (idx < 0) return [];
-    shelf = smartShelves[idx];
-    isSmart = true;
-    listLen = smartShelves.length;
-  }
-  const isHidden = !!shelf?.hidden;
-  let isCollapsed = false;
-  try { isCollapsed = (globalThis as any).localStorage?.getItem?.(`ds-collapsed-${shelfId}`) === "1"; } catch {}
-
-  const lbl = (key: string, fallback: string): string => {
-    try { const v = i18n.t(key as any); return (typeof v === "string" && v && v !== key) ? v : fallback; } catch { return fallback; }
-  };
-
-  // A menu action mutates settings, which re-renders the home and drops
-  // gamepad focus — Steam then defaults it to the first shelf. Pin the card
-  // the menu was opened on and let the focus-restore loop put it back after
-  // the re-render (the loop's confirmation poll also defends against Steam's
-  // post-render first-card grab).
-  const preserveFocus = () => {
-    if (focusedAppId > 0) {
-      try { saveFocusTarget(focusedAppId, shelfId); beginFocusRestoreLoop(); } catch {}
-    }
-  };
-  const item = (key: string, label: string, onSelected: () => void, disabled?: boolean) =>
-    R.createElement(dfl.MenuItem, { key, onSelected: () => { onSelected(); preserveFocus(); }, disabled }, label);
-
-  const src: any = shelf?.source;
-  const isOnline = src?.type === "wishlist" || src?.type === "store";
-  const isRandomOrSmart =
-    isSmart ||
-    isOnline ||
-    src?.type === "smart" ||
-    shelf?.sort === "random" ||
-    (src?.type === "filter" && src?.filter?.sort === "random");
-
-  // Toggle a boolean flag on the shelf and persist immediately.
-  const toggleFlag = (key: string) => {
-    const s = getCurrentSettings();
-    if (!s) return;
-    const next = !shelf[key];
-    if (isSmart) {
-      const updated = (s.smartShelves ?? []).map((sh: any) =>
-        sh.id === shelfId ? { ...sh, [key]: next } : sh,
-      );
-      void saveSettings({ ...s, smartShelves: updated });
-    } else {
-      void saveSettings(patchShelfInSettings(s, shelfId, { [key]: next } as any));
-    }
-  };
-
-  // Prefix checked items with a checkmark so the current state is visible.
-  const checked = (flag: boolean, label: string) => (flag ? `✓ ${label}` : label);
-
-  // ── Management submenu ────────────────────────────────────────────────
-  // Decoration is NOT here — it lives as its own submenu at the
-  // Prateleira level (same visual rank as Display / Visual), built
-  // below. Smart shelves skip the decoration submenu entirely
-  // (resolver is mode-driven, not slot-driven).
-  const mgmt = [
-    item("ds-edit",      lbl("editShelf", "Edit"),      () => dispatchShelfModal("edit", shelfId)),
-    item("ds-duplicate", lbl("duplicateShelf", "Duplicate"), () => {
-      void duplicateShelfById(shelfId, lbl("copySuffix", "(Copy)"));
-    }),
-    item(
-      "ds-collapse",
-      isCollapsed ? lbl("expand_shelf", "Expand shelf") : lbl("collapse_shelf", "Collapse shelf"),
-      () => setShelfCollapsed(shelfId, !isCollapsed),
-    ),
-    item(
-      "ds-hide",
-      isHidden ? lbl("show_shelf", "Show shelf") : lbl("hide_shelf", "Hide shelf"),
-      () => { void toggleShelfHiddenById(shelfId); },
-    ),
-    item("ds-move-up",   lbl("move_up", "Move up"),     () => { void moveShelfById(shelfId, -1); }, idx <= 0),
-    item("ds-move-down", lbl("move_down", "Move down"),  () => { void moveShelfById(shelfId, 1); }, idx >= listLen - 1),
-    ...(isRandomOrSmart ? [
-      item("ds-refresh", isOnline ? lbl("refresh_cache", "Refresh cache") : lbl("refresh", "Refresh"), () => {
-        try {
-          if (isSmart || src?.type === "smart") invalidateSmartShelfCache(shelfId);
-          else if (isOnline) clearOnlineShelfCache();
-          else invalidateRandomSortCache(shelfId);
-          triggerShelfRefresh({ manual: true, shelfId });
-        } catch {}
-      }),
-    ] : []),
-    item("ds-delete", lbl("deleteShelf", "Delete"), () => dispatchShelfModal("delete", shelfId)),
-  ];
-
-  // ── Display submenu ───────────────────────────────────────────────────
-  const display = [
-    item("ds-d-title",     checked(!!shelf.hideShelfTitle,    lbl("hide_shelf_title",    "Hide shelf title")),    () => toggleFlag("hideShelfTitle")),
-    item("ds-d-names",     checked(!!shelf.hideGameNames,     lbl("hide_game_name",      "Hide game names")),     () => toggleFlag("hideGameNames")),
-    item("ds-d-status",    checked(!!shelf.hideStatusLine,    lbl("hide_status_line",    "Hide status line")),    () => toggleFlag("hideStatusLine")),
-    item("ds-d-badge",     checked(!!shelf.hideNewBadge,      lbl("hide_new_badge",      "Hide new badge")),      () => toggleFlag("hideNewBadge")),
-    item("ds-d-discount",  checked(!!shelf.hideDiscountBadge, lbl("hide_discount_badge", "Hide discount badge")), () => toggleFlag("hideDiscountBadge")),
-    item("ds-d-compat",    checked(!!shelf.hideCompatIcons,   lbl("hide_compat_icons",   "Hide compat icons")),   () => toggleFlag("hideCompatIcons")),
-    item("ds-d-nsbadge",   checked(!!shelf.hideNonSteamBadge, lbl("hide_non_steam_badge","Hide non-Steam badge")),() => toggleFlag("hideNonSteamBadge")),
-    item("ds-d-install",   checked(!!shelf.hideInstallIndicator, lbl("hide_install_indicator","Hide install indicator")), () => toggleFlag("hideInstallIndicator")),
-    item("ds-d-seemore",   checked(!!shelf.hideSeeMore,       lbl("hide_see_more_card",  "Hide \"See more\"")),   () => toggleFlag("hideSeeMore")),
-    item("ds-d-refresh",   checked(!!shelf.hideRefreshCard,   lbl("hide_refresh_card",   "Hide refresh card")),   () => toggleFlag("hideRefreshCard")),
-  ];
-
-  // ── Visual submenu ────────────────────────────────────────────────────
-  const visual = [
-    item("ds-v-native",    checked(!!shelf.matchNativeSize,   lbl("match_native_size",   "Match native size")),   () => toggleFlag("matchNativeSize")),
-    item("ds-v-hiFirst",   checked(!!shelf.highlightFirst,    lbl("highlight_first",     "Highlight first card")),() => toggleFlag("highlightFirst")),
-    item("ds-v-hiAll",     checked(!!shelf.highlightAll,      lbl("highlight_all",       "Highlight all cards")), () => toggleFlag("highlightAll")),
-    item("ds-v-hero",      checked(!!(shelf as any).heroEnabled, lbl("hero_enabled_label","Enable hero art")),     () => toggleFlag("heroEnabled")),
-  ];
-
-  const group = (key: string, label: string, ...children: any[]) =>
-    R.createElement(dfl.MenuGroup, { key, label }, ...children);
-
-  // ── Sort direction toggle (direct item inside "Prateleira") ───────────
-  const isReversed = !!shelf.sortReverse;
-  const sortLabel = isReversed
-    ? checked(true, lbl("sort_descending", "Sort: descending"))
-    : lbl("sort_ascending", "Sort: ascending");
-  const sortToggle = item("ds-sort-dir", sortLabel, () => toggleFlag("sortReverse"));
-
-  // ── Card-level actions (base level, no submenu) ───────────────────────
-  // These apply to the specific focused card, not the shelf as a whole.
-  const cardActions: any[] = [];
-  if (focusedAppId > 0) {
-    // A card is effectively highlighted if it's individually in highlightedAppIds,
-    // OR if it's at index 0 and highlightFirst is on (shelf-level setting),
-    // OR if highlightAll is on. Reflect all three in the menu so the user
-    // can see the active state and toggle it per-card without confusion.
-    const inHighlightedIds = (shelf.highlightedAppIds ?? []).includes(focusedAppId);
-    const isFirstCard = _activeCardIndexForMenu === 0;
-    const highlightedViaFirst = isFirstCard && !!shelf.highlightFirst;
-    const highlightedViaAll = !!shelf.highlightAll;
-    const highlighted = inHighlightedIds || highlightedViaFirst || highlightedViaAll;
-    cardActions.push(item(
-      "ds-card-highlight",
-      highlighted
-        ? checked(true, lbl("remove_highlight", "Remove highlight"))
-        : lbl("highlight_this", "Highlight this game"),
-      () => {
-        const s = getCurrentSettings();
-        if (!s) return;
-        // When removing highlight: if the highlight came from highlightAll/
-        // highlightFirst (shelf-level), turn those off. If it came from
-        // highlightedAppIds (per-card), remove from that list.
-        if (highlighted) {
-          const patch: Record<string, any> = {};
-          if (highlightedViaAll) patch.highlightAll = false;
-          if (highlightedViaFirst) patch.highlightFirst = false;
-          if (inHighlightedIds) patch.highlightedAppIds = (shelf.highlightedAppIds ?? []).filter((id: number) => id !== focusedAppId);
-          if (isSmart) {
-            const updated = (s.smartShelves ?? []).map((sh: any) =>
-              sh.id === shelfId ? { ...sh, ...patch } : sh,
-            );
-            void saveSettings({ ...s, smartShelves: updated });
-          } else {
-            void saveSettings(patchShelfInSettings(s, shelfId, patch));
-          }
-        } else {
-          const ids: number[] = shelf.highlightedAppIds ?? [];
-          const next = [...ids, focusedAppId];
-          if (isSmart) {
-            const updated = (s.smartShelves ?? []).map((sh: any) =>
-              sh.id === shelfId ? { ...sh, highlightedAppIds: next } : sh,
-            );
-            void saveSettings({ ...s, smartShelves: updated });
-          } else {
-            void saveSettings(patchShelfInSettings(s, shelfId, { highlightedAppIds: next }));
-          }
-        }
-      },
-    ));
-
-    const hiddenFromShelf = (shelf.hiddenAppIds ?? []).includes(focusedAppId);
-    cardActions.push(item(
-      "ds-card-hide",
-      hiddenFromShelf
-        ? lbl("show_in_shelf", "Show in shelf")
-        : lbl("hide_from_shelf", "Hide from shelf"),
-      () => {
-        const s = getCurrentSettings();
-        if (!s) return;
-        const ids: number[] = shelf.hiddenAppIds ?? [];
-        const next = hiddenFromShelf ? ids.filter((id) => id !== focusedAppId) : [...ids, focusedAppId];
-        if (isSmart) {
-          const updated = (s.smartShelves ?? []).map((sh: any) =>
-            sh.id === shelfId ? { ...sh, hiddenAppIds: next } : sh,
-          );
-          void saveSettings({ ...s, smartShelves: updated });
-        } else {
-          void saveSettings(patchShelfInSettings(s, shelfId, { hiddenAppIds: next }));
-        }
-      },
-    ));
-
-    // ── Add-to-shelf submenu ────────────────────────────────────────────
-    // Append-target shelves (excludes current, already-in, at-limit,
-    // 50+ manual entries). Selecting flips to manual sort + appends.
-    const ABSOLUTE_MAX = 50;
-    const candidateShelves: any[] = (settings.shelves ?? []).filter((sh: any) => {
-      if (sh.id === shelfId) return false;
-      const manual: number[] = sh.manualOrder ?? [];
-      if (manual.includes(focusedAppId)) return false;
-      const cap = Math.min(typeof sh.limit === "number" ? sh.limit : ABSOLUTE_MAX, ABSOLUTE_MAX);
-      if (manual.length >= cap) return false;
-      return true;
-    });
-    if (candidateShelves.length > 0) {
-      const addItems = candidateShelves.map((sh: any) => item(
-        `ds-card-add-${sh.id}`,
-        sh.title ?? sh.id,
-        () => {
-          const s = getCurrentSettings();
-          if (!s) return;
-          const tgt: any = (s.shelves ?? []).find((row: any) => row.id === sh.id);
-          if (!tgt) return;
-          const manual: number[] = tgt.manualOrder ?? [];
-          const wasManual = tgt.sort === "manual";
-          const patch: Record<string, any> = {
-            sort: "manual",
-            sortReverse: false,
-            manualOrder: [...manual, focusedAppId],
-          };
-          if (!wasManual) {
-            // Preserve the current natural order so the rest of the row
-            // doesn't reshuffle when we engage manual mode.
-            patch.manualBaseSort = typeof tgt.sort === "string" ? tgt.sort : "alphabetical";
-          }
-          void saveSettings(patchShelfInSettings(s, sh.id, patch));
-        },
-      ));
-      cardActions.push(group("ds-card-add-shelf", lbl("menu_add_to_shelf", "Add to shelf"), ...addItems));
-    }
-
-    // ── Remove-from-shelf submenu ──────────────────────────────────────
-    // Lists every regular shelf whose manualOrder currently contains
-    // this appid (excluding the shelf the menu is rooted on — the
-    // existing "Hide from shelf" entry handles the current shelf).
-    // Selecting an entry strips the appid from that shelf's manualOrder.
-    const removableShelves: any[] = (settings.shelves ?? []).filter((sh: any) => {
-      if (sh.id === shelfId) return false;
-      const manual: number[] = sh.manualOrder ?? [];
-      return manual.includes(focusedAppId);
-    });
-    if (removableShelves.length > 0) {
-      const rmItems = removableShelves.map((sh: any) => item(
-        `ds-card-rm-${sh.id}`,
-        sh.title ?? sh.id,
-        () => {
-          const s = getCurrentSettings();
-          if (!s) return;
-          const tgt: any = (s.shelves ?? []).find((row: any) => row.id === sh.id);
-          if (!tgt) return;
-          const manual: number[] = tgt.manualOrder ?? [];
-          void saveSettings(patchShelfInSettings(s, sh.id, {
-            manualOrder: manual.filter((id) => id !== focusedAppId),
-          }));
-        },
-      ));
-      cardActions.push(group("ds-card-remove-shelf", lbl("menu_remove_from_shelf", "Remove from shelf"), ...rmItems));
-    }
-  }
-
-  // Decoration shortcut — plain item at the Prateleira level (no group
-  // wrapper), same visual rank as Display / Visual but selecting it
-  // jumps straight to the modal's Decoration tab instead of opening a
-  // single-item submenu.
-  const decoration = isSmart ? [] : [
-    item("ds-decoration", lbl("menu_decoration", "Decoration"), () => dispatchShelfModal("edit", shelfId, { initialTab: "decoration" })),
-  ];
-
-  // Card-level items sit at the same level as "Prateleira", before it.
-  // The "Prateleira" group contains only shelf-scoped submenus.
-  return [
-    ...cardActions,
-    group(
-      "ds-shelf-root", lbl("menu_shelf", "Shelf"),
-      sortToggle,
-      group("ds-mgmt",    lbl("menu_management", "Management"), ...mgmt),
-      group("ds-display", lbl("menu_display",    "Display"),    ...display),
-      group("ds-visual",  lbl("menu_visual",     "Visual"),     ...visual),
-      ...decoration,
-    ),
-  ];
 }
 
 /**
@@ -1635,91 +1340,6 @@ export function extractAppContextMenu(): boolean {
   }
 
   return false;
-}
-
-/**
- * Public wrapper for `buildDeckShelvesMenuItems` — used by `Shelf.tsx`
- * (online shelf card menu) to get the standard DS menu structure without
- * opening a native Steam menu. Returns an empty array when the shelf can't
- * be resolved.
- */
-export function buildShelfContextMenu(shelfId: string, appid: number, dfl: any, R: any): any[] {
-  return buildDeckShelvesMenuItems(shelfId, dfl, R, appid);
-}
-
-/**
- * Library-card Add/Remove-to-shelf injection. Emits up to two
- * `MenuGroup` submenus (one per action) with one `MenuItem` per
- * eligible shelf. Uses the STATIC `@decky/ui` imports so the component
- * references match what Steam's renderer expects exactly — at runtime
- * `window.DFL.MenuGroup` and the import resolve to the same instance,
- * but importing keeps the reference stable across plugin reloads and
- * Decky bundle updates (vs reading from a global that might drift).
- *
- * Returns `[]` when no eligible shelves exist on either action.
- */
-export function buildLibraryAddToShelfItems(appid: number, _dfl: any, R: any): any[] {
-  if (!R?.createElement) return [];
-  if (!appid) return [];
-  const s = getCurrentSettings?.();
-  if (!s) return [];
-  const ABSOLUTE_MAX = 50;
-  const eligible: any[] = (s.shelves ?? []).filter((sh: any) => {
-    const manual: number[] = sh.manualOrder ?? [];
-    if (manual.includes(appid)) return false;
-    const cap = Math.min(typeof sh.limit === "number" ? sh.limit : ABSOLUTE_MAX, ABSOLUTE_MAX);
-    if (manual.length >= cap) return false;
-    return true;
-  });
-  const removable: any[] = (s.shelves ?? []).filter((sh: any) => (sh.manualOrder ?? []).includes(appid));
-  if (!eligible.length && !removable.length) return [];
-  const lblFn = (key: string, fallback: string): string => {
-    try { const v = i18n.t(key as any); return (typeof v === "string" && v && v !== key) ? v : fallback; } catch { return fallback; }
-  };
-  const mkItem = (key: string, label: string, onSelected: () => void) =>
-    R.createElement(DeckyMenuItem, { key, onSelected }, label);
-
-  const groups: any[] = [];
-  if (eligible.length > 0) {
-    const addChildren = eligible.map((sh: any) => mkItem(
-      `ds-lib-add-${sh.id}`,
-      sh.title ?? sh.id,
-      () => {
-        const cur = getCurrentSettings();
-        if (!cur) return;
-        const tgt: any = (cur.shelves ?? []).find((row: any) => row.id === sh.id);
-        if (!tgt) return;
-        const manual: number[] = tgt.manualOrder ?? [];
-        const wasManual = tgt.sort === "manual";
-        const patch: Record<string, any> = {
-          sort: "manual",
-          sortReverse: false,
-          manualOrder: [...manual, appid],
-        };
-        if (!wasManual) patch.manualBaseSort = typeof tgt.sort === "string" ? tgt.sort : "alphabetical";
-        void saveSettings(patchShelfInSettings(cur, sh.id, patch));
-      },
-    ));
-    groups.push(R.createElement(DeckyMenuGroup, { key: "ds-lib-add-shelf", label: lblFn("menu_add_to_shelf", "Add to shelf") }, ...addChildren));
-  }
-  if (removable.length > 0) {
-    const rmChildren = removable.map((sh: any) => mkItem(
-      `ds-lib-rm-${sh.id}`,
-      sh.title ?? sh.id,
-      () => {
-        const cur = getCurrentSettings();
-        if (!cur) return;
-        const tgt: any = (cur.shelves ?? []).find((row: any) => row.id === sh.id);
-        if (!tgt) return;
-        const manual: number[] = tgt.manualOrder ?? [];
-        void saveSettings(patchShelfInSettings(cur, sh.id, {
-          manualOrder: manual.filter((id) => id !== appid),
-        }));
-      },
-    ));
-    groups.push(R.createElement(DeckyMenuGroup, { key: "ds-lib-remove-shelf", label: lblFn("menu_remove_from_shelf", "Remove from shelf") }, ...rmChildren));
-  }
-  return groups;
 }
 
 export function showGameMenu(appid: number, shelfId?: string): void {
